refactor: User 서비스 인터페이스 생성 및 User 생성자에 role 추가
