+++ src/main/java/team03/monew/service/user/AuthService.java
@@ -1,35 +1,11 @@
 package team03.monew.service.user;
 
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 import team03.monew.dto.user.UserDto;
 import team03.monew.dto.user.UserLoginRequest;
-import team03.monew.entity.user.User;
-import team03.monew.mapper.user.UserMapper;
-import team03.monew.repository.user.UserRepository;
-import team03.monew.util.exception.user.InvalidException;
-import team03.monew.util.exception.user.UserNotFoundException;
 
-@Slf4j
-@Service
-@RequiredArgsConstructor
-@Transactional
-public class AuthService {
+public interface AuthService {
 
-  private final UserRepository userRepository;
-  private final UserMapper userMapper;
+  // 로그인
+  UserDto login(UserLoginRequest request);
 
-  public UserDto login(UserLoginRequest request) {
-    log.debug("로그인 시작: email={}", request.email());
-    User user = userRepository.findByEmail(request.email())
-        .orElseThrow(() -> UserNotFoundException.withEmail(request.email()));
-
-    if (!user.getPassword().equals(request.password())) {
-      throw InvalidException.wrongPassword();
-    }
-    log.info("로그인 완료: userId={}, email={}", user.getId(), user.getEmail());
-    return userMapper.toDto(user);
-  }
 }
